Start abstracting the functions of the test classes out into a Tests class.
Use START/END_OF_TEST_JSON instead of framework-specific strings in output.

This doesn't work yet.
